#3 -- Rename schema
Renamed response to resolution in schema for relay local state.

Response is too vague. The prior Response component was for
handling a response message, not creating a response. Changed
that Response component to Mutation.

Resolution record is created when resolve the GraphQL query
in the Resolve component, Resolution name fits.